+++ src/orm/entities/Conversation.ts
@@ -5,10 +5,15 @@ import { ConversationUser } from './ConversationUser';
 
 @Entity()
 export class Conversation extends Base {
-
-  @OneToMany(() => ConversationUser, conversationUser => conversationUser.conversation)
+  @OneToMany(
+    () => ConversationUser,
+    conversationUser => conversationUser.conversation
+  )
   public conversationUser?: ConversationUser[];
 
-  @OneToMany(() => Message, (message) => message.conversation)
-  messages: Message[];
+  @OneToMany(
+    () => Message,
+    message => message.conversation
+  )
+  messages?: Message[];
 }