+++ src/orm/entities/ConversationUser.ts
@@ -5,19 +5,23 @@ import { Conversation } from './Conversation';
 
 @Entity('conversation_users_user')
 export class ConversationUser extends Base {
-
-  @Column({name: 'user_id'})
+  @Column({ name: 'user_id' })
   userId: string;
 
-  @Column({name: 'conversation_id' })
+  @Column({ name: 'conversation_id' })
   conversationId: string;
 
-  @ManyToOne(() => User, (user) => user.conversationUser)
+  @ManyToOne(
+    () => User,
+    user => user.conversationUser
+  )
   @JoinColumn({ name: 'user_id' })
-  public user: User
+  public user?: User;
 
-  @ManyToOne(() => Conversation, (conversation) => conversation.conversationUser)
+  @ManyToOne(
+    () => Conversation,
+    conversation => conversation.conversationUser
+  )
   @JoinColumn({ name: 'conversation_id' })
-  public conversation: Conversation
-
+  public conversation?: Conversation;
 }