+++ src/orm/entities/User.ts
@@ -5,13 +5,18 @@ import { ConversationUser } from './ConversationUser';
 
 @Entity()
 export class User extends Base {
-
   @Column()
   name: string;
 
-  @OneToMany(() => Message, (message) => message.user)
-  messages: Message[];
+  @OneToMany(
+    () => Message,
+    message => message.user
+  )
+  messages?: Message[];
 
-  @OneToMany(() => ConversationUser, conversationUser => conversationUser.user)
+  @OneToMany(
+    () => ConversationUser,
+    conversationUser => conversationUser.user
+  )
   public conversationUser?: ConversationUser[];
 }